+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ jobs:
         node: [ 16.x, 18.x ]
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
 
       - name: Warming up (node)
         uses: actions/setup-node@v3
