+++ .github/workflows/release.yml
@@ -10,7 +10,7 @@ jobs:
 
     if: github.repository_owner == 'jspaste'
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
 
       - name: Warming up (node)
         uses: actions/setup-node@v3
