+++ lib/core/JSPError.ts
